Remove action log dependency; remove defaction.
First of all, we don't want to couple the interruption code to the
action log service, which is optional. `defaction` as written was
knitting those two things together. In the absence of being able to find
a good way to keep interruption and action logging completely separate
within simulant itself, defaction was converted to an example people can
use when writing their sim test applications.